+++ Controls-demo/dropdown_new/Button/HeaderTemplate/Index.ts
@@ -1,23 +1,9 @@
 import {Control, TemplateFunction} from 'UI/Base';
 import controlTemplate = require('wml!Controls-demo/dropdown_new/Button/HeaderTemplate/Index');
-import {Memory} from 'Types/source';
 
-class HeaderContentTemplate extends Control {
+export default class extends Control {
     protected _template: TemplateFunction = controlTemplate;
-    protected _source: Memory;
-
-    protected _beforeMount(): void {
-        this._source = new Memory({
-            keyProperty: 'key',
-            data: [
-                {key: 1, title: 'Ярославль'},
-                {key: 2, title: 'Москва'},
-                {key: 3, title: 'Санкт-Петербург'}
-            ]
-        });
-    }
 
     static _theme: string[] = ['Controls/Classes'];
-    static _styles: string[] = ['Controls-demo/Controls-demo', 'Controls-demo/dropdown_new/Button/Index'];
+    static _styles: string[] = ['Controls-demo/Controls-demo'];
 }
-export default HeaderContentTemplate;
