+++ Controls-demo/dropdown_new/Button/HeaderTemplate/Index.wml
@@ -1,15 +1,18 @@
-<div class="controlsDemo__flexRow">
-   <div class="controlsDemo__wrapper controlsDemo-Dropdown_absolute">
-      <Controls.dropdown:Button
-              keyProperty="key"
-              displayProperty="title"
-              caption="{[Выберите город]}"
-              source="{{_source}}"
-              attr:class="controlsDemo-menuButton">
-         <ws:headerTemplate>
-            <ws:partial template="Controls/dropdown:HeaderTemplate" caption="{[Выберите город]}" />
-         </ws:headerTemplate>
-      </Controls.dropdown:Button>
-      <div class="controls-text-label_theme-{{_options.theme}} controlsDemo-Dropdown_label">Открытие меню вверх</div>
+<div class="controlsDemo__wrapper">
+   <div class="controlsDemo__cell">
+      <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
+                               state="/Controls-demo/app/:app"
+                               app="{{ 'Controls-demo/dropdown_new/Button/HeaderTemplate/Simple/Index' }}">
+         <a href="{{content.href}}">Меню с шапкой</a>
+      </Router.router:Reference>
+      <Controls-demo.dropdown_new.Button.HeaderTemplate.Simple.Index/>
+   </div>
+   <div class="controlsDemo__cell">
+      <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
+                               state="/Controls-demo/app/:app"
+                               app="{{ 'Controls-demo/dropdown_new/Button/HeaderTemplate/IconSize/Index' }}">
+         <a href="{{content.href}}">Меню с разным размером иконок в шапке и в пунктах</a>
+      </Router.router:Reference>
+      <Controls-demo.dropdown_new.Button.HeaderTemplate.IconSize.Index/>
    </div>
 </div>
