+++ Controls-demo/dropdown_new/Button/Index.less
@@ -1,3 +1,7 @@
+.controlsDemo-Dropdown_button_scroll {
+   height: 100%;
+}
+
 .controlsDemo-Dropdown_absolute {
    position: absolute;
    bottom: 0;
@@ -24,4 +28,8 @@
    &_bottom {
       bottom: 15px;
    }
+}
+
+.ControlsDemo-Page {
+   height: 100%;
 }
