+++ Controls/_dropdown/ComboBox.ts
@@ -32,6 +32,7 @@ const getPropValue = Utils.object.getPropertyValue.bind(Utils);
  *
  * @remark
  * Полезные ссылки:
+ * * {@link /materials/Controls-demo/app/Controls-demo%2Fdropdown_new%2FCombobox%2FIndex демо-пример}
  * * {@link /doc/platform/developmentapl/interface-development/controls/dropdown-menu/combobox/ руководство разработчика}
  * * {@link https://github.com/saby/wasaby-controls/blob/rc-20.4000/Controls-default-theme/aliases/_dropdown.less переменные тем оформления dropdown}
  * * {@link https://github.com/saby/wasaby-controls/blob/rc-20.4000/Controls-default-theme/aliases/_dropdownPopup.less переменные тем оформления dropdownPopup}
